+++ lib/conditional_switch.dart
@@ -11,7 +11,8 @@ class ConditionalSwitch {
   /// when one of the keys matches the value returns by [valueBuilder],
   /// the corresponding `Widget` builder will be used.
   /// - [fallbackBuilder] is a function which returns a `Widget`,
-  ///  it is used when none of the keys in [caseBuilders] matches the value returns by [valueBuilder].
+  ///  it is used when none of the keys in [caseBuilders]matches
+  /// the value returns by [valueBuilder].
   static Widget single({
     @required dynamic Function() valueBuilder,
     @required Map<dynamic, Widget Function()> caseBuilders,
@@ -32,7 +33,8 @@ class ConditionalSwitch {
   /// when one of the keys matches the value returns by [valueBuilder],
   /// the corresponding `List<Widget>` builder will be used.
   /// - [fallbackBuilder] is a function which returns a `List<Widget>`,
-  ///  it is used when none of the keys in [caseBuilders] matches the value returns by [valueBuilder].
+  ///  it is used when none of the keys in [caseBuilders] matches
+  /// the value returns by [valueBuilder].
   static List<Widget> list({
     @required dynamic Function() valueBuilder,
     @required Map<dynamic, List<Widget> Function()> caseBuilders,